smgl-keyring-diff: do all git commands in a branch
This way we don't lose any of the commits that aren't pushed yet. Changes that
weren't pushed were lost due to "git rebase -q origin master". Due to the fact
that now all changes are made in a branch master isn't touched and that temp
branch can simply be discarded (deleted) at the end.